requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-openai
langchain_community
langchain-google-genai
langchain-groq
langgraph
openai
google-generativeai
langchain_groq
chromadb
faiss-cpu
tiktoken
pymupdf
python-docx
pandas
unstructured
requests
streamlit
azure-ai-inference
azure-core
python-multipart
sentence-transformers
langchain_openai
bcrypt<4.1.0
pydantic[email]
passlib[bcrypt]
python-jose
jwt
PyJWT

#Evaluation
ragas
mlflow
evaluate
datasets
bert-score
absl-py
rouge_score